+++ src/Booker.Backend/Controllers/ServiceController.cs
@@ -3,17 +3,17 @@
 [Route("api/[controller]")]
 [ApiController]
 [Authorize]
-public class ServiceController(IServiceService serviceService) : ControllerBase
+public class ServiceController(IServiceService serviceService, IValidatorService validatorService) : ControllerBase
 {
     [HttpGet]
     [Route("{userId}")]
     [ProducesResponseType(typeof(List<ServiceDto>), StatusCodes.Status200OK)]
-    public async Task<ActionResult<List<ServiceDto>>> GetServicesForUser(
-        [FromRoute] string userId,
-        CancellationToken cancellationToken
-    )
+    public async Task<ActionResult<List<ServiceDto>>> GetServicesForUser(CancellationToken cancellationToken)
     {
-        var services = await serviceService.GetServicesForUser(userId, cancellationToken);
+        var services = await serviceService.GetServicesForUser(
+            User.FindFirstValue(JwtRegisteredClaimNames.Sub)!,
+            cancellationToken
+        );
         return Ok(services);
     }
 
@@ -39,11 +39,22 @@ public class ServiceController(IServiceService serviceService) : ControllerBase
     [HttpPost, Authorize(Roles = "Admin, Provider")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
     public async Task<IActionResult> AddService(
         [FromBody] EditServiceRequest newService,
         CancellationToken cancellationToken
     )
     {
+        var userIsOwner = await validatorService.ValidateCalendarOwnership(
+            newService.CalendarId,
+            User.FindFirstValue(JwtRegisteredClaimNames.Sub)!
+        );
+
+        if (!userIsOwner)
+        {
+            return Forbid("The user does not have permission to edit this calendar.");
+        }
+
         var result = await serviceService.AddService(newService, cancellationToken);
 
         if (result is not null)
@@ -57,11 +68,22 @@ public class ServiceController(IServiceService serviceService) : ControllerBase
     [HttpPut, Authorize(Roles = "Admin, Provider")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
     public async Task<IActionResult> UpdateService(
         [FromBody] EditServiceRequest updatedService,
         CancellationToken cancellationToken
     )
     {
+        var userIsOwner = await validatorService.ValidateCalendarOwnership(
+            updatedService.CalendarId,
+            User.FindFirstValue(JwtRegisteredClaimNames.Sub)!
+        );
+
+        if (!userIsOwner)
+        {
+            return Forbid("The user does not have permission to edit this calendar.");
+        }
+
         var result = await serviceService.UpdateService(updatedService, cancellationToken);
 
         if (result is not null)
@@ -75,8 +97,19 @@ public class ServiceController(IServiceService serviceService) : ControllerBase
     [HttpDelete("{id:int}"), Authorize(Roles = "Admin, Provider")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(typeof(string), StatusCodes.Status403Forbidden)]
     public async Task<IActionResult> DeleteService(int id, CancellationToken cancellationToken)
     {
+        var userIsOwner = await validatorService.ValidateCalendarOwnership(
+            id,
+            User.FindFirstValue(JwtRegisteredClaimNames.Sub)!
+        );
+
+        if (!userIsOwner)
+        {
+            return Forbid("The user does not have permission to edit this calendar.");
+        }
+
         var errorMessage = await serviceService.DeleteService(id, cancellationToken);
 
         if (errorMessage is not null)
